icebreaker: remove internal oscillator and PLL, and add --freq to PNRFLAGS
The internal oscillator and PLL were removed to make the code more
readable for students. How to instance the PLL and make the processor
run faster will be explained in the FAQ.

Removed the PNR flag that allowed the timing to fail, we don't want
that! Instead we now have --freq 13 (push a bit harsher than just 12
MHz, but it should also work with 12)